Move user registraiton component files to src/UserPopup
