+++ src/UserPopup/ReduxUserRegistration.js
